# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "My API"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": ["http"],
  "paths": {
...
    "/api/receipt": {
      "get": {
        "tags": ["Receipt"],
        "summary": "Get all receipts",
        "operationId": "getReceipts",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/Receipt"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error"
          }
        }
      },
      "post": {
        "tags": ["Receipt"],
        "summary": "Create a new receipt",
        "operationId": "createReceipt",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/definitions/Receipt"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/definitions/Receipt"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    }
  },
  "definitions": {
...
    "Receipt": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "pictureUrl": {
          "type": "string"
        },
        "loyalPercent": {
          "type": "number"
        },
        "address": {
          "type": "string"
        },
        "workers": {
          "type": "array",
          "items": {
            "type": "string",
            "format": "uuid"
          }
        },
        "userOwner": {
          "type": "string",
          "format": "uuid"
        },
        "business": {
          "type": "string",
          "format": "uuid"
        },
        "receipts": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Receipt"
          }
        }
      },
      "required": [
        "name",
        "userOwner",
        "business",
        "receipts"
      ]
    }
  }
}